+++ cx-portal/src/components/shared/basic/ReleaseProcess/AppPage/index.tsx
@@ -39,13 +39,14 @@ import {
   increment,
 } from 'features/appManagement/slice'
 import {
+  DocumentTypeId,
   useFetchAppStatusQuery,
   useUpdateappMutation,
   useUpdateDocumentUploadMutation,
 } from 'features/appManagement/apiSlice'
 import { setAppStatus } from 'features/appManagement/actions'
 import { Dropzone } from 'components/shared/basic/Dropzone'
-import SnackbarNotificationWithButtons from '../SnackbarNotificationWithButtons'
+import SnackbarNotificationWithButtons from '../components/SnackbarNotificationWithButtons'
 import { ConnectorFormInputField } from '../components/ConnectorFormInputField'
 import ReleaseStepHeader from '../components/ReleaseStepHeader'
 import ProviderConnectorField from '../components/ProviderConnectorField'
@@ -111,7 +112,7 @@ export default function AppPage() {
   })
 
   useEffect(() => {
-    dispatch(setAppStatus(fetchAppStatus))
+    if (fetchAppStatus) dispatch(setAppStatus(fetchAppStatus))
   }, [dispatch, fetchAppStatus])
 
   const uploadAppContractValue = getValues().uploadAppContract
@@ -209,7 +210,7 @@ export default function AppPage() {
     if (value && !('status' in value)) {
       setFileStatus('uploadAppContract', UploadStatus.UPLOADING)
 
-      uploadDocumentApi(appId, 'APP_CONTRACT', value)
+      uploadDocumentApi(appId, DocumentTypeId.APP_CONTRACT, value)
         .then(() =>
           setFileStatus('uploadAppContract', UploadStatus.UPLOAD_SUCCESS)
         )
@@ -226,7 +227,7 @@ export default function AppPage() {
     if (value && !('status' in value)) {
       setFileStatus('uploadDataPrerequisits', UploadStatus.UPLOADING)
 
-      uploadDocumentApi(appId, 'ADDITIONAL_DETAILS', value)
+      uploadDocumentApi(appId, DocumentTypeId.ADDITIONAL_DETAILS, value)
         .then(() =>
           setFileStatus('uploadDataPrerequisits', UploadStatus.UPLOAD_SUCCESS)
         )
@@ -243,7 +244,7 @@ export default function AppPage() {
     if (value && !('status' in value)) {
       setFileStatus('uploadTechnicalGuide', UploadStatus.UPLOADING)
 
-      uploadDocumentApi(appId, 'APP_TECHNICAL_INFORMATION', value)
+      uploadDocumentApi(appId, DocumentTypeId.APP_TECHNICAL_INFORMATION, value)
         .then(() =>
           setFileStatus('uploadTechnicalGuide', UploadStatus.UPLOAD_SUCCESS)
         )
@@ -269,7 +270,7 @@ export default function AppPage() {
 
       for (let fileIndex = 0; fileIndex < value.length; fileIndex++) {
         setFileStatus(fileIndex, UploadStatus.UPLOADING)
-        uploadDocumentApi(appId, 'APP_IMAGE', value[fileIndex])
+        uploadDocumentApi(appId, DocumentTypeId.APP_IMAGE, value[fileIndex])
           .then(() => setFileStatus(fileIndex, UploadStatus.UPLOAD_SUCCESS))
           .catch(() => setFileStatus(fileIndex, UploadStatus.UPLOAD_ERROR))
       }
@@ -278,7 +279,7 @@ export default function AppPage() {
 
   const uploadDocumentApi = async (
     appId: string,
-    documentTypeId: string,
+    documentTypeId: DocumentTypeId,
     file: any
   ) => {
     const data = {
@@ -343,7 +344,7 @@ export default function AppPage() {
   }
 
   const onBackIconClick = () => {
-    dispatch(setAppStatus(fetchAppStatus))
+    if (fetchAppStatus) dispatch(setAppStatus(fetchAppStatus))
     dispatch(decrement())
   }
 
@@ -566,8 +567,8 @@ export default function AppPage() {
           title: t('content.apprelease.appReleaseForm.error.title'),
           description: t('content.apprelease.appReleaseForm.error.message'),
         }}
-        setPageNotification={setAppPageNotification}
-        setPageSnackbar={setAppPageSnackbar}
+        setPageNotification={() => setAppPageNotification(false)}
+        setPageSnackbar={() => setAppPageSnackbar(false)}
         onBackIconClick={onBackIconClick}
         onSave={handleSubmit((data) => onAppPageSubmit(data, 'save'))}
         onSaveAndProceed={handleSubmit((data) =>