+++ cx-portal/src/components/shared/basic/ReleaseProcess/components/ReleaseProcessWrapper.tsx
@@ -39,6 +39,7 @@ import OfferCard from '../OfferCard'
 import { currentActiveStep } from 'features/appManagement/slice'
 import { useSelector } from 'react-redux'
 import OfferPage from '../OfferPage'
+import OfferContractAndConsent from '../OfferContractAndConsent'
 
 interface ReleaseProcessWrapperType {
   headerTitle: string
@@ -68,7 +69,7 @@ export default function ReleaseProcessWrapper({
   onAppsOverviewClick,
 }: ReleaseProcessWrapperType) {
   const [showSubmitPage, setShowSubmitPage] = useState(false)
-  let activePage: any = useSelector(currentActiveStep)
+  let activePage: number = useSelector(currentActiveStep)
   useEffect(() => {
     activeStep()
     window.scrollTo(0, 0)
@@ -87,7 +88,7 @@ export default function ReleaseProcessWrapper({
     } else {
       if (activePage === 1) return <OfferCard />
       else if (activePage === 2) return <OfferPage />
-      else if (activePage === 3) return <ContractAndConsent />
+      else if (activePage === 3) return <OfferContractAndConsent />
       else if (activePage === 4)
         return <ValidateAndPublish showSubmitPage={setShowSubmitPage} />
     }