+++ cx-portal/src/features/appManagement/slice.ts
@@ -24,6 +24,7 @@ import {
   SearchInputState,
   name,
   AppStatusDataState,
+  ServiceStatusDataState,
 } from './types'
 import { RootState } from 'features/store'
 
@@ -53,6 +54,14 @@ const managementSlice = createSlice({
       ...state,
       appStatusData: action.payload,
     }),
+    setServiceStatusData: (state, action) => ({
+      ...state,
+      serviceStatusData: action.payload,
+    }),
+    setServiceId: (state, action) => ({
+      ...state,
+      serviceId: action.payload,
+    }),
   },
 })
 
@@ -65,10 +74,22 @@ export const currentActiveStep = (state: RootState): any =>
 export const appIdSelector = (state: RootState): string =>
   state.management.appId
 
+export const serviceIdSelector = (state: RootState): string =>
+  state.management.serviceId
+
 export const appStatusDataSelector = (state: RootState): AppStatusDataState =>
   state.management.appStatusData
 
-export const { increment, decrement, setCurrentActiveStep, setApplicationId } =
-  managementSlice.actions
+export const serviceStatusDataSelector = (
+  state: RootState
+): ServiceStatusDataState => state.management.serviceStatusData
+
+export const {
+  increment,
+  decrement,
+  setCurrentActiveStep,
+  setApplicationId,
+  setServiceId,
+} = managementSlice.actions
 
 export default managementSlice